refactor: kumoy.sprite
