+++ kumoy/sprite/sprite_packer.py
