+++ kumoy/sprite/symbol_collector.py
