+++ kumoy/sprite/symbol_normalizer.py
