+++ kumoy/sprite/uploader.py
